Introduce models module
Up to now, models have been loaded from the inference module. Now, the get their own module.
Additionally, a FiestaModel base class has been introduced and the Surrogate classes have been renamed to make clear that these are machine-learning surrogates.
