+++ docs/quickstart/analytical_models.rst
@@ -38,7 +38,7 @@ source-frame times (days):
 .. code:: python
 
     import jax.numpy as jnp
-    from fiesta.inference.analytical_models import OneComponentKilonovaModel
+    from fiesta.models.analytical_models import OneComponentKilonovaModel
 
     model = OneComponentKilonovaModel(
         filters=["bessellb", "bessellv", "bessellr"],
@@ -72,8 +72,8 @@ where the afterglow is a trained surrogate:
 
 .. code:: python
 
-    from fiesta.inference.lightcurve_model import AfterglowFlux, CombinedSurrogate
-    from fiesta.inference.analytical_models import OneComponentKilonovaModel
+    from fiesta.models.surrogate_models import AfterglowFlux, CombinedSurrogate
+    from fiesta.models.analytical_models import OneComponentKilonovaModel
 
     afterglow = AfterglowFlux(
         name="afgpy_gaussian_CVAE",
@@ -644,7 +644,7 @@ To add a new analytical model, subclass ``AnalyticalModel`` and implement
 .. code:: python
 
     import jax.numpy as jnp
-    from fiesta.inference.analytical_models import AnalyticalModel
+    from fiesta.models.analytical_models import AnalyticalModel
 
     class MyModel(AnalyticalModel):
         parameter_names = ["log10_luminosity", "log10_radius"]
