+++ src/fiesta/inference/__init__.py
@@ -1,6 +1,5 @@
 from .fiesta import Fiesta
 from .likelihood import EMLikelihood, FluxLikelihood, LikelihoodBase
-from .lightcurve_model import FluxModel, LightcurveModel, CombinedSurrogate, SurrogateModel
 from .plot import corner_plot, LightcurvePlotter
 
 __all__ = [
