+++ src/fiesta/models/__init__.py
@@ -1,0 +1,8 @@
+from .surrogate_models import FluxSurrogate
+from .combined_model import CombinedModel
+
+
+__all__ = [
+    name for name in dir()
+    if not name.startswith("_")
+]
