+++ src/fiesta/models/analytical_models/base.py
