+++ surrogates/GRB/afgpy_gaussian_CVAE/train_afterglowpy_gaussian.py
@@ -3,7 +3,6 @@ import matplotlib.pyplot as plt
 import h5py
 
 from fiesta.train.FluxTrainer import CVAETrainer, DataManager
-from fiesta.inference.lightcurve_model import AfterglowFlux
 from fiesta.train.neuralnets import NeuralnetConfig
 
 #############
@@ -69,9 +68,4 @@ trainer.save()
 print("Producing example lightcurve . . .")
 FILTERS = ["radio-3GHz", "X-ray-1keV", "radio-6GHz", "bessellv"]
 
-lc_model = AfterglowFlux(name,
-                         directory=outdir, 
-                         filters=FILTERS)
-
-
-trainer.plot_example_lc(lc_model)
+trainer.plot_example_lc(FILTERS)
