+++ surrogates/GRB/pbag_tophat_CVAE/train_pyblastafterglow_tophat.py
@@ -3,7 +3,6 @@ import matplotlib.pyplot as plt
 import h5py
 
 from fiesta.train.FluxTrainer import CVAETrainer
-from fiesta.inference.lightcurve_model import AfterglowFlux
 from fiesta.train.neuralnets import NeuralnetConfig
 
 #############
@@ -67,7 +66,4 @@ trainer.save()
 print("Producing example lightcurve . . .")
 FILTERS = ["radio-3GHz", "X-ray-1keV", "radio-6GHz", "bessellv"]
 
-lc_model = AfterglowFlux(name,
-                         directory="./model",
-                         filters = FILTERS)
-trainer.plot_example_lc(lc_model)
+trainer.plot_example_lc(FILTERS)
