+++ surrogates/KN/Brethauer2025_CVAE/train_Brethauer2025.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import h5py
 
 from fiesta.train.FluxTrainer import CVAETrainer
-from fiesta.inference.lightcurve_model import FluxModel
+from fiesta.models.surrogate_models import FluxSurrogate
 from fiesta.train.neuralnets import NeuralnetConfig
 from fiesta.train.Benchmarker import Benchmarker
 
@@ -70,20 +70,23 @@ trainer.save()
 print("Producing example lightcurve . . .")
 
 FILTERS = ["ps1::y", "besselli", "bessellv", "bessellux"]
-lc_model = FluxModel(name,
-                     directory=outdir, 
+trainer.plot_example_lc(FILTERS)
+
+
+####################
+### BENCHMARKING ###
+####################
+
+
+lc_model = FluxSurrogate(name,
+                     directory=outdir,
                      filters=FILTERS)
 
-trainer.plot_example_lc(lc_model)
-
-
 for metric_name in ["L2", "Linf"]:
-
     benchmarker = Benchmarker(
                     model = lc_model,
                     data_file =file,
                     metric_name = metric_name
                     )
-    
     benchmarker.benchmark()
     benchmarker.plot_lightcurves_mismatch()
