+++ surrogates/KN/Bu2025_MLP/benchmark_Bu2025.py
@@ -1,5 +1,5 @@
 from fiesta.train.Benchmarker import Benchmarker
-from fiesta.inference.lightcurve_model import AfterglowFlux
+from fiesta.models.surrogate_models import AfterglowFlux
