+++ surrogates/KN/Bu2025_MLP/train_Bu2025.py
@@ -3,7 +3,6 @@ import matplotlib.pyplot as plt
 import h5py
 
 from fiesta.train.FluxTrainer import PCATrainer
-from fiesta.inference.lightcurve_model import FluxModel
 from fiesta.train.neuralnets import NeuralnetConfig
 
 #############
@@ -69,9 +68,5 @@ trainer.save()
 print("Producing example lightcurve . . .")
 
 FILTERS = ["ps1::y", "besselli", "bessellv", "bessellux"]
-lc_model = FluxModel(name,
-                     directory=outdir, 
-                     filters=FILTERS)
-
-trainer.plot_example_lc(lc_model)
+trainer.plot_example_lc(FILTERS)
