+++ surrogates/KN/Bu2026_CVAE/train_Bu2026.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import h5py
 
 from fiesta.train.FluxTrainer import CVAETrainer
-from fiesta.inference.lightcurve_model import FluxModel
+from fiesta.models.surrogate_models import FluxSurrogate
 from fiesta.train.neuralnets import NeuralnetConfig
 from fiesta.train.Benchmarker import Benchmarker
 
@@ -65,11 +65,17 @@ trainer.save()
 print("Producing example lightcurve . . .")
 
 FILTERS = ["2massj", "besselli", "bessellv", "bessellux"]
-lc_model = FluxModel(name,
-                     directory=outdir, 
-                     filters=FILTERS)
+trainer.plot_example_lc(FILTERS)
 
-trainer.plot_example_lc(lc_model)
+
+####################
+### BENCHMARKING ###
+####################
+
+
+lc_model = FluxSurrogate(name,
+                     directory=outdir,
+                     filters=FILTERS)
 
 for metric_name in ["L2", "Linf"]:
     benchmarker = Benchmarker(
